Fix CI build errors, warnings, and test gaps
- Fix GCC -Wmisleading-indentation in 11 parameter load functions (add braces)
- Fix -Wreorder in particle_oscillator.h and xy_morph_pad.h initializer lists
- Fix -Wpedantic extra semicolons in adsr_display.h
- Fix -Wsign-compare in controller.cpp (cast enum to int32_t)
- Fix -Wformat-truncation in step_pattern_editor.h (widen buffer)
- Fix missing #include <iomanip> in mod_ring_indicator.h (macOS build)
- Split mod_source_colors.h into mod_matrix_types.h (no VSTGUI dep) + color header
  so processor doesn't depend on VSTGUI and tests can compile without it
- Fix ruinae_tests linker errors by adding vstgui_uidescription/vstgui_support
  with entry point stubs for test executables
- Fix ADSR display tests to expect 1px minimum segment width (not 15%)
- Enforce 1px minimum segment width in ADSR layout computation
- Remove unused enforceMinimumFractions and kMinSegmentWidthFraction
- Add shared_tests build+run to CI on all three platforms

+++ plugins/ruinae/src/controller/controller.cpp
@@ -510,7 +510,7 @@ VSTGUI::CView* Controller::verifyView(
     auto* control = dynamic_cast<VSTGUI::CControl*>(view);
     if (control) {
         auto tag = control->getTag();
-        if (tag >= kActionPresetAllTag && tag <= kActionEuclideanRegenTag) {
+        if (tag >= static_cast<int32_t>(kActionPresetAllTag) && tag <= static_cast<int32_t>(kActionEuclideanRegenTag)) {
             control->registerControlListener(this);
         }
     }
